server/requirements.txt
celery==5.1.2
Django==3.2.6
django-celery-beat==2.2.1
django-cors-headers==3.7.0
django-filter==2.4.0
django-simple-history==3.0.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
django-bulk-update-or-create==0.3.0
drf-yasg==1.20.0
psutil==5.8.0
pymysql==1.0.2